Alimentazione: proteine/carboidrati/grassi accettano una cifra decimale
Prima erano numeri interi, ma un'etichetta nutrizionale spesso riporta
valori come 23.5g. I tre campi passano da PositiveIntegerField a
DecimalField(decimal_places=1) - le kcal restano intere (raramente
hanno senso frazionate). Aggiornati i campi input (step 0.1) e la
visualizzazione (arrotonda a una cifra, senza mostrare ".0" se il
valore e' intero). Gli obiettivi restano numeri interi per ora, dato
che di solito sono target tondi scelti a mano - dimmi se vuoi che
diventino decimali anche quelli.

Migration 0033.

+++ tracker/models.py
@@ -129,9 +129,9 @@ class MacroEntry(models.Model):
     # scomposizione in macro di quel pasto.
     utente = models.ForeignKey(User, on_delete=models.CASCADE, related_name='macro_entries')
     kcal = models.PositiveIntegerField()
-    proteine_g = models.PositiveIntegerField(default=0)
-    carboidrati_g = models.PositiveIntegerField(default=0)
-    grassi_g = models.PositiveIntegerField(default=0)
+    proteine_g = models.DecimalField(max_digits=5, decimal_places=1, default=0)
+    carboidrati_g = models.DecimalField(max_digits=5, decimal_places=1, default=0)
+    grassi_g = models.DecimalField(max_digits=5, decimal_places=1, default=0)
     nota = models.CharField(max_length=100, blank=True, default='')
     data = models.DateField(default=timezone.now)
     creato_il = models.DateTimeField(default=timezone.now)
